Task 3

+++ DesignPrinciples_Calculator/src/main/java/com/epam/Calculate.java
@@ -2,7 +2,7 @@ package com.epam;
 
 public interface Calculate {
 	public int add(int a , int b);
-	public int mul(int a , int b);
+	public double mul(int a , int b);
 	public int div(int a , int b);
 	public int sub(int a , int b);
